requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
google-cloud-storage